# contracts/openapi/primehr-v1.yaml
openapi: 3.0.3
info:
  title: ISOFT PRIME-HRM API
  version: 2.2.0-phase-2.2
  description: Phase 1 competency foundation and Phase 2.1 position-profile draft administration. All operations require an ISOFT employee JWT and server-resolved agency scope. Position-profile commands use the independent primehr.position-profile permission and validate authoritative targets through Administrative.
servers:
  - url: /api/primehr/v1
security:
  - bearerAuth: []
paths:
...
  /admin/position-profiles/{id}:
    get:
      operationId: getPositionProfile
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses:
        '200': {$ref: '#/components/responses/PositionProfileRecord'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      operationId: updatePositionProfileDraft
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {$ref: '#/components/requestBodies/UpdatePositionProfile'}
      responses:
        '200': {$ref: '#/components/responses/PositionProfileRecord'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/PositionTargetUnavailable'}
  /admin/position-profiles/{id}/versions:
    post:
      operationId: createPositionProfileSuccessorDraft
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {$ref: '#/components/requestBodies/Transition'}
      responses:
        '201': {$ref: '#/components/responses/PositionProfileRecord'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/PositionTargetUnavailable'}
  /admin/position-profiles/{id}/archive:
    post:
      operationId: archivePositionProfileDraft
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {$ref: '#/components/requestBodies/Transition'}
      responses:
        '200': {$ref: '#/components/responses/PositionProfileRecord'}
        '409': {$ref: '#/components/responses/Conflict'}
  /admin/position-profiles/{id}/requirements:
    post:
      operationId: addPositionProfileRequirement
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {$ref: '#/components/requestBodies/CreatePositionRequirement'}
      responses:
        '201': {$ref: '#/components/responses/PositionProfileRecord'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {$ref: '#/components/responses/Conflict'}
  /admin/position-profiles/{id}/requirements/{requirementId}:
    put:
      operationId: updatePositionProfileRequirement
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/RequirementId'
      requestBody: {$ref: '#/components/requestBodies/UpdatePositionRequirement'}
      responses:
        '200': {$ref: '#/components/responses/PositionProfileRecord'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {$ref: '#/components/responses/Conflict'}
  /admin/position-profiles/{id}/requirements/{requirementId}/archive:
    post:
      operationId: archivePositionProfileRequirement
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/RequirementId'
      requestBody: {$ref: '#/components/requestBodies/RequirementTransition'}
      responses:
        '200': {$ref: '#/components/responses/PositionProfileRecord'}
        '409': {$ref: '#/components/responses/Conflict'}
  /admin/position-profiles/{id}/submit:
    post:
      operationId: submitPositionProfile
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {$ref: '#/components/requestBodies/SubmitPositionProfile'}
      responses:
        '200': {$ref: '#/components/responses/PositionProfileRecord'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/PositionTargetUnavailable'}
  /admin/position-profiles/{id}/return:
    post:
      operationId: returnPositionProfileSubmission
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {$ref: '#/components/requestBodies/Transition'}
      responses:
        '200': {$ref: '#/components/responses/PositionProfileRecord'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
  /admin/position-profiles/{id}/approve:
    post:
      operationId: approvePositionProfile
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {$ref: '#/components/requestBodies/ApprovePositionProfile'}
      responses:
        '200': {$ref: '#/components/responses/PositionProfileRecord'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/PositionTargetUnavailable'}
  /admin/position-profiles/{id}/audit-events:
    get:
      operationId: listPositionProfileAuditEvents
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {$ref: '#/components/responses/AuditEventPage'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /admin/position-profiles/resolve:
    get:
      operationId: resolveEffectivePositionProfile
      parameters:
        - {name: jobPositionId, in: query, required: true, schema: {type: integer, format: int64, minimum: 1}}
        - {name: plantillaId, in: query, required: false, schema: {type: integer, format: int64, minimum: 1}}
        - {name: asOf, in: query, required: true, schema: {type: string, format: date}}
      responses:
        '200': {$ref: '#/components/responses/PositionProfileResolution'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /admin/position-profiles/compare:
    get:
      operationId: comparePositionProfileVersions
      parameters:
        - {name: leftProfileId, in: query, required: true, schema: {type: string, maxLength: 36}}
        - {name: rightProfileId, in: query, required: true, schema: {type: string, maxLength: 36}}
      responses:
        '200': {$ref: '#/components/responses/PositionProfileComparison'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page: {name: page, in: query, schema: {type: integer, minimum: 0, default: 0}}
    Size: {name: size, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
...
    Id: {name: id, in: path, required: true, schema: {type: string, maxLength: 36}}
...
    RequirementId: {name: requirementId, in: path, required: true, schema: {type: string, maxLength: 36}}
...
  requestBodies:
...
    Transition: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TransitionRequest'}}}}
...
    UpdatePositionProfile: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UpdatePositionProfileRequest'}}}}
    CreatePositionRequirement: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreatePositionRequirementRequest'}}}}
    UpdatePositionRequirement: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UpdatePositionRequirementRequest'}}}}
    RequirementTransition: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PositionRequirementTransitionRequest'}}}}
    SubmitPositionProfile: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SubmitPositionProfileRequest'}}}}
    ApprovePositionProfile: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ApprovePositionProfileRequest'}}}}
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
...
    Forbidden:
      description: Authenticated identity lacks Competency read permission or a trusted agency context
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    NotFound:
      description: Agency-scoped definition was not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    Conflict: {description: Optimistic conflict or illegal lifecycle transition, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
...
    PositionTargetUnavailable: {description: The authoritative Administrative position target could not be verified, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
...
    PositionProfileRecord: {description: Versioned position competency profile, content: {application/json: {schema: {$ref: '#/components/schemas/PositionProfile'}}}}
...
    PositionProfileResolution: {description: Effective profile selected by Plantilla-then-Job Position precedence, content: {application/json: {schema: {$ref: '#/components/schemas/PositionProfileResolution'}}}}
    PositionProfileComparison: {description: Exact-version requirement comparison, content: {application/json: {schema: {$ref: '#/components/schemas/PositionProfileComparison'}}}}
    AuditEventPage: {description: Page of append-only position-profile audit events, content: {application/json: {schema: {$ref: '#/components/schemas/AuditEventPage'}}}}
  schemas:
...
    TransitionRequest:
      type: object
      required: [recordVersion, reason]
      properties: {recordVersion: {type: integer, format: int64, minimum: 0}, reason: {type: string, minLength: 1, maxLength: 1000}}
...
    PositionTargetType: {type: string, enum: [JOB_POSITION, PLANTILLA]}
    RequirementClassification: {type: string, enum: [MANDATORY, DESIRABLE]}
...
    UpdatePositionProfileRequest:
      type: object
      required: [name, recordVersion]
      properties:
        name: {type: string, minLength: 1, maxLength: 200}
        description: {type: string, maxLength: 2000, nullable: true}
        effectiveFrom: {type: string, format: date, nullable: true}
        effectiveTo: {type: string, format: date, nullable: true}
        recordVersion: {type: integer, format: int64, minimum: 0}
    CreatePositionRequirementRequest:
      type: object
      required: [competencyVersionId, requiredProficiencyLevelId, classification, displayOrder, profileRecordVersion]
      properties:
        competencyVersionId: {type: string, maxLength: 36}
        requiredProficiencyLevelId: {type: string, maxLength: 36}
        classification: {$ref: '#/components/schemas/RequirementClassification'}
        criticalityCode: {type: string, maxLength: 50, nullable: true}
        remarks: {type: string, maxLength: 2000, nullable: true}
        displayOrder: {type: integer, minimum: 0}
        profileRecordVersion: {type: integer, format: int64, minimum: 0}
    UpdatePositionRequirementRequest:
      type: object
      required: [requiredProficiencyLevelId, classification, displayOrder, recordVersion, profileRecordVersion]
      properties:
        requiredProficiencyLevelId: {type: string, maxLength: 36}
        classification: {$ref: '#/components/schemas/RequirementClassification'}
        criticalityCode: {type: string, maxLength: 50, nullable: true}
        remarks: {type: string, maxLength: 2000, nullable: true}
        displayOrder: {type: integer, minimum: 0}
        recordVersion: {type: integer, format: int64, minimum: 0}
        profileRecordVersion: {type: integer, format: int64, minimum: 0}
    PositionRequirementTransitionRequest:
      type: object
      required: [recordVersion, profileRecordVersion, reason]
      properties:
        recordVersion: {type: integer, format: int64, minimum: 0}
        profileRecordVersion: {type: integer, format: int64, minimum: 0}
        reason: {type: string, minLength: 1, maxLength: 1000}
    SubmitPositionProfileRequest:
      type: object
      required: [recordVersion]
      properties:
        recordVersion: {type: integer, format: int64, minimum: 0}
    ApprovePositionProfileRequest:
      type: object
      required: [recordVersion]
      properties:
        recordVersion: {type: integer, format: int64, minimum: 0}
        reason: {type: string, maxLength: 1000, nullable: true, description: Required for administrator self-approval.}
    PositionTargetSnapshot:
      type: object
      required: [type, targetId, jobPositionId, jobPositionName, sourceFingerprint, capturedAt]
      properties:
        type: {$ref: '#/components/schemas/PositionTargetType'}
        targetId: {type: integer, format: int64}
        jobPositionId: {type: integer, format: int64}
        jobPositionName: {type: string, maxLength: 200}
        salaryGrade: {type: integer, format: int64, nullable: true}
        salaryStep: {type: integer, format: int64, nullable: true}
        plantillaId: {type: integer, format: int64, nullable: true}
        plantillaName: {type: string, maxLength: 200, nullable: true}
        sourceFingerprint: {type: string, minLength: 64, maxLength: 64}
        capturedAt: {type: string, format: date-time}
    PositionRequirement:
      type: object
      required: [id, competencyVersionId, competencyCode, competencyName, competencyDefinitionVersion, requiredProficiencyLevelId, requiredProficiencyLevelCode, requiredProficiencyLevelLabel, classification, active, displayOrder, recordVersion]
      properties:
        id: {type: string, maxLength: 36}
        competencyVersionId: {type: string, maxLength: 36}
        competencyCode: {type: string, maxLength: 50}
        competencyName: {type: string, maxLength: 200}
        competencyDefinitionVersion: {type: integer, minimum: 1}
        requiredProficiencyLevelId: {type: string, maxLength: 36}
        requiredProficiencyLevelCode: {type: string, maxLength: 50}
        requiredProficiencyLevelLabel: {type: string, maxLength: 150}
        classification: {$ref: '#/components/schemas/RequirementClassification'}
        criticalityCode: {type: string, maxLength: 50, nullable: true}
        remarks: {type: string, maxLength: 2000, nullable: true}
        active: {type: boolean}
        displayOrder: {type: integer, minimum: 0}
        recordVersion: {type: integer, format: int64, minimum: 0}
    PositionProfile:
      type: object
      required: [id, name, status, definitionVersion, recordVersion, contentRevision, targetSnapshot, requirements]
      properties:
        id: {type: string, maxLength: 36}
        name: {type: string, maxLength: 200}
        description: {type: string, maxLength: 2000, nullable: true}
        status: {type: string, enum: [DRAFT, SUBMITTED, ACTIVE, ARCHIVED]}
        definitionVersion: {type: integer, minimum: 1}
        supersedesId: {type: string, maxLength: 36, nullable: true}
        effectiveFrom: {type: string, format: date, nullable: true}
        effectiveTo: {type: string, format: date, nullable: true}
        recordVersion: {type: integer, format: int64, minimum: 0}
        contentRevision: {type: integer, format: int64, minimum: 0}
        submittedBy: {type: string, maxLength: 100, nullable: true}
        submittedAt: {type: string, format: date-time, nullable: true}
        approvedBy: {type: string, maxLength: 100, nullable: true}
        approvedAt: {type: string, format: date-time, nullable: true}
        targetSnapshot: {$ref: '#/components/schemas/PositionTargetSnapshot'}
        requirements: {type: array, items: {$ref: '#/components/schemas/PositionRequirement'} }
...
    PositionProfileResolution:
      type: object
      required: [asOf, resolvedBy, profile]
      properties:
        asOf: {type: string, format: date}
        resolvedBy: {$ref: '#/components/schemas/PositionTargetType'}
        profile: {$ref: '#/components/schemas/PositionProfile'}
    PositionProfileComparisonItem:
      type: object
      required: [change, competencyVersionId, competencyCode, competencyName]
      properties:
        change: {type: string, enum: [ADDED, REMOVED, CHANGED, UNCHANGED]}
        competencyVersionId: {type: string, maxLength: 36}
        competencyCode: {type: string, maxLength: 50}
        competencyName: {type: string, maxLength: 200}
        leftLevelId: {type: string, maxLength: 36, nullable: true}
        leftLevelLabel: {type: string, maxLength: 150, nullable: true}
        leftClassification: {allOf: [{$ref: '#/components/schemas/RequirementClassification'}], nullable: true}
        leftCriticalityCode: {type: string, maxLength: 50, nullable: true}
        rightLevelId: {type: string, maxLength: 36, nullable: true}
        rightLevelLabel: {type: string, maxLength: 150, nullable: true}
        rightClassification: {allOf: [{$ref: '#/components/schemas/RequirementClassification'}], nullable: true}
        rightCriticalityCode: {type: string, maxLength: 50, nullable: true}
    PositionProfileComparison:
      type: object
      required: [leftProfileId, leftDefinitionVersion, rightProfileId, rightDefinitionVersion, added, removed, changed, unchanged, items]
      properties:
        leftProfileId: {type: string, maxLength: 36}
        leftDefinitionVersion: {type: integer, minimum: 1}
        rightProfileId: {type: string, maxLength: 36}
        rightDefinitionVersion: {type: integer, minimum: 1}
        added: {type: integer, minimum: 0}
        removed: {type: integer, minimum: 0}
        changed: {type: integer, minimum: 0}
        unchanged: {type: integer, minimum: 0}
        items: {type: array, items: {$ref: '#/components/schemas/PositionProfileComparisonItem'}}
    AuditEvent:
      type: object
      required: [id, actor, action, aggregateType, aggregateId, occurredAt]
      properties:
        id: {type: string, maxLength: 36}
        actor: {type: string, maxLength: 100}
        action: {type: string, maxLength: 50}
        aggregateType: {type: string, maxLength: 100}
        aggregateId: {type: string, maxLength: 36}
        businessVersion: {type: integer, nullable: true}
        recordVersion: {type: integer, format: int64, nullable: true}
        occurredAt: {type: string, format: date-time}
        previousState: {type: string, nullable: true}
        newState: {type: string, nullable: true}
        reason: {type: string, maxLength: 1000, nullable: true}
        correlationId: {type: string, maxLength: 100, nullable: true}
    AuditEventPage:
      type: object
      required: [content, page, size, totalElements, totalPages, first, last]
      properties:
        content: {type: array, items: {$ref: '#/components/schemas/AuditEvent'}}
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1}
        totalElements: {type: integer, format: int64, minimum: 0}
        totalPages: {type: integer, minimum: 0}
        first: {type: boolean}
        last: {type: boolean}
...
    ApiError:
      type: object
      required: [timestamp, status, message, path, details]
      properties:
        timestamp: {type: string, format: date-time}
        status: {type: integer}
        message: {type: string}
        path: {type: string}
        details:
          type: array
          items: {type: string}